Refactor command to their own files and add tests
